requirements.txt
colorama==0.4.6
colorlog==6.7.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
protobuf==3.20.3
PyYAML==6.0
